Fix Color in ListWidget and bad /ip packet
Color in listwidget is now preserved if a new peer connects while a
message is pending.  Bad /ip packets were being sent that contained
ipaddresses with identical hostnames.  Don't send a hostname/ip combo in
that packet if they are the same.  That combo is usually the person
requesting the IPs to begin with!  FIXME:change that so its not a race
condition