TCPRead change, prevent partial reads
Prevent partial reads for large messages (>2000).  All messages send are
now screened for a length.  This is then used to modify the watermark to
prevent an early call to a callback.  Bad lengths should just cause a
drain of the entire buffer, more work is needed here to prevent this
drain from happening 4bytes at a time forever until a valid read is
made.  WIP